rmk: отказываемся от BigDecimal для сумм — абсолютная точность не требуется

+++ app/src/main/java/com/micdm/smsgraphs/data/Operation.java
@@ -1,6 +1,5 @@
 package com.micdm.smsgraphs.data;
 
-import java.math.BigDecimal;
 import java.util.Calendar;
 
 public class Operation {
@@ -8,9 +7,9 @@ public class Operation {
     public final int id;
     public final Target target;
     public final Calendar created;
-    public final BigDecimal amount;
+    public final int amount;
 
-    public Operation(int id, Target target, Calendar created, BigDecimal amount) {
+    public Operation(int id, Target target, Calendar created, int amount) {
         this.id = id;
         this.target = target;
         this.created = created;
